Match Timer
Time limit with end conditions are added. There is a null reference after the timer goes to 0 and keeps going to -1, -2 etc. I removed one of the 'Scene1' files which I believe was causing issues when playing the prototype. Besides the null reference, I think the code should work fine.