Match end conditions added
I added draw and win/lose conditions for the game. There are a few null references once the game starts but the code works fine. There is a build issue when returning to the main menu on draw, however it seems to be a build settings issue that isn't particularly serious.